@@ -1,25 +1,27 @@
 name: retrofit_generator
 description: retrofit generator is an dio client generator using source_gen and inspired by Chopper and Retrofit.
-version: 5.0.0
-
 homepage: https://mings.in/retrofit.dart/
 repository: https://github.com/trevorwang/retrofit.dart/
-
+topics:
+  - build-runner
+  - codegen
+  - api
+version: 7.0.8
 environment:
-  sdk: '>=2.17.0 <3.0.0'
+  sdk: '>=2.19.0 <4.0.0'
 
 dependencies:
-  analyzer: '>=4.7.0 <6.0.0'
+  analyzer: '>=5.13.0 <7.0.0'
   build: ^2.3.1
   built_collection: ^5.1.1
-  code_builder: ^4.3.0
-  dart_style: ^2.2.4
+  code_builder: ^4.4.0
+  dart_style: ^2.3.0
   dio: ^5.0.0
   retrofit: ^4.0.1
-  source_gen: ^1.2.5
-  tuple: ^2.0.0
+  source_gen: ^1.3.0
+  tuple: ^2.0.1
 
 dev_dependencies:
   lints: ^2.0.1
-  source_gen_test: ^1.0.4
-  test: ^1.21.4
+  source_gen_test: ^1.0.5
+  test: ^1.24.0
