@@ -18,8 +18,13 @@ Map<String, dynamic> _$UserToJson(User instance) => <String, dynamic>{
 // RetrofitGenerator
 // **************************************************************************
 
+// ignore_for_file: unnecessary_brace_in_string_interps,no_leading_underscores_for_local_identifiers
+
 class _RestClient implements RestClient {
-  _RestClient(this._dio, {this.baseUrl}) {
+  _RestClient(
+    this._dio, {
+    this.baseUrl,
+  }) {
     baseUrl ??= 'http://baidu.com';
   }
 
@@ -33,17 +38,17 @@ class _RestClient implements RestClient {
     final queryParameters = <String, dynamic>{};
     queryParameters.removeWhere((k, v) => v == null);
     final _headers = <String, dynamic>{};
-    final _data = <String, dynamic>{};
+    final Map<String, dynamic>? _data = null;
     final newOptions = newRequestOptions(options);
     newOptions.extra.addAll(_extra);
     newOptions.headers.addAll(_dio.options.headers);
     newOptions.headers.addAll(_headers);
     final _result = await _dio.fetch<List<dynamic>>(newOptions.copyWith(
-        method: 'GET',
-        baseUrl: baseUrl ?? _dio.options.baseUrl,
-        queryParameters: queryParameters,
-        path: '/tags')
-      ..data = _data);
+      method: 'GET',
+      baseUrl: baseUrl ?? _dio.options.baseUrl,
+      queryParameters: queryParameters,
+      path: '/tags',
+    )..data = _data);
     final value = _result.data!.cast<String>();
     return value;
   }
@@ -54,17 +59,17 @@ class _RestClient implements RestClient {
     final queryParameters = <String, dynamic>{};
     queryParameters.removeWhere((k, v) => v == null);
     final _headers = <String, dynamic>{};
-    final _data = <String, dynamic>{};
+    final Map<String, dynamic>? _data = null;
     final newOptions = newRequestOptions(options);
     newOptions.extra.addAll(_extra);
     newOptions.headers.addAll(_dio.options.headers);
     newOptions.headers.addAll(_headers);
     final _result = await _dio.fetch<List<dynamic>>(newOptions.copyWith(
-        method: 'GET',
-        baseUrl: baseUrl ?? _dio.options.baseUrl,
-        queryParameters: queryParameters,
-        path: '/tagsNullable')
-      ..data = _data);
+      method: 'GET',
+      baseUrl: baseUrl ?? _dio.options.baseUrl,
+      queryParameters: queryParameters,
+      path: '/tagsNullable',
+    )..data = _data);
     final value = _result.data?.cast<String>();
     return value;
   }
@@ -75,17 +80,17 @@ class _RestClient implements RestClient {
     final queryParameters = <String, dynamic>{};
     queryParameters.removeWhere((k, v) => v == null);
     final _headers = <String, dynamic>{};
-    final _data = <String, dynamic>{};
+    final Map<String, dynamic>? _data = null;
     final newOptions = newRequestOptions(options);
     newOptions.extra.addAll(_extra);
     newOptions.headers.addAll(_dio.options.headers);
     newOptions.headers.addAll(_headers);
     final _result = await _dio.fetch<Map<String, dynamic>>(newOptions.copyWith(
-        method: 'GET',
-        baseUrl: baseUrl ?? _dio.options.baseUrl,
-        queryParameters: queryParameters,
-        path: '/tagByKey')
-      ..data = _data);
+      method: 'GET',
+      baseUrl: baseUrl ?? _dio.options.baseUrl,
+      queryParameters: queryParameters,
+      path: '/tagByKey',
+    )..data = _data);
     final value = _result.data!.cast<String, String>();
     return value;
   }
@@ -96,40 +101,43 @@ class _RestClient implements RestClient {
     final queryParameters = <String, dynamic>{};
     queryParameters.removeWhere((k, v) => v == null);
     final _headers = <String, dynamic>{};
-    final _data = <String, dynamic>{};
+    final Map<String, dynamic>? _data = null;
     final newOptions = newRequestOptions(options);
     newOptions.extra.addAll(_extra);
     newOptions.headers.addAll(_dio.options.headers);
     newOptions.headers.addAll(_headers);
     final _result = await _dio.fetch<Map<String, dynamic>>(newOptions.copyWith(
-        method: 'GET',
-        baseUrl: baseUrl ?? _dio.options.baseUrl,
-        queryParameters: queryParameters,
-        path: '/tagByKeyNullable')
-      ..data = _data);
+      method: 'GET',
+      baseUrl: baseUrl ?? _dio.options.baseUrl,
+      queryParameters: queryParameters,
+      path: '/tagByKeyNullable',
+    )..data = _data);
     final value = _result.data?.cast<String, String>();
     return value;
   }
 
   @override
-  Future<String> getTag({options, required contentType}) async {
+  Future<String> getTag({
+    options,
+    required contentType,
+  }) async {
     const _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     queryParameters.removeWhere((k, v) => v == null);
     final _headers = <String, dynamic>{r'Content-Type': contentType};
     _headers.removeWhere((k, v) => v == null);
-    final _data = <String, dynamic>{};
+    final Map<String, dynamic>? _data = null;
     final newOptions = newRequestOptions(options);
     newOptions.extra.addAll(_extra);
     newOptions.headers.addAll(_dio.options.headers);
     newOptions.headers.addAll(_headers);
     final _result = await _dio.fetch<String>(newOptions.copyWith(
-        method: 'GET',
-        contentType: contentType,
-        baseUrl: baseUrl ?? _dio.options.baseUrl,
-        queryParameters: queryParameters,
-        path: '/tag')
-      ..data = _data);
+      method: 'GET',
+      contentType: contentType,
+      baseUrl: baseUrl ?? _dio.options.baseUrl,
+      queryParameters: queryParameters,
+      path: '/tag',
+    )..data = _data);
     final value = _result.data!;
     return value;
   }
@@ -140,17 +148,17 @@ class _RestClient implements RestClient {
     final queryParameters = <String, dynamic>{};
     queryParameters.removeWhere((k, v) => v == null);
     final _headers = <String, dynamic>{};
-    final _data = <String, dynamic>{};
+    final Map<String, dynamic>? _data = null;
     final newOptions = newRequestOptions(options);
     newOptions.extra.addAll(_extra);
     newOptions.headers.addAll(_dio.options.headers);
     newOptions.headers.addAll(_headers);
     final _result = await _dio.fetch<String>(newOptions.copyWith(
-        method: 'GET',
-        baseUrl: baseUrl ?? _dio.options.baseUrl,
-        queryParameters: queryParameters,
-        path: '/tagNullable')
-      ..data = _data);
+      method: 'GET',
+      baseUrl: baseUrl ?? _dio.options.baseUrl,
+      queryParameters: queryParameters,
+      path: '/tagNullable',
+    )..data = _data);
     final value = _result.data;
     return value;
   }
@@ -161,19 +169,21 @@ class _RestClient implements RestClient {
     final queryParameters = <String, dynamic>{};
     queryParameters.removeWhere((k, v) => v == null);
     final _headers = <String, dynamic>{};
-    final _data = <String, dynamic>{};
+    final Map<String, dynamic>? _data = null;
     final newOptions = newRequestOptions(options);
     newOptions.extra.addAll(_extra);
     newOptions.headers.addAll(_dio.options.headers);
     newOptions.headers.addAll(_headers);
     final _result = await _dio.fetch<List<dynamic>>(newOptions.copyWith(
-        method: 'GET',
-        baseUrl: baseUrl ?? _dio.options.baseUrl,
-        queryParameters: queryParameters,
-        path: '/users')
-      ..data = _data);
+      method: 'GET',
+      baseUrl: baseUrl ?? _dio.options.baseUrl,
+      queryParameters: queryParameters,
+      path: '/users',
+    )..data = _data);
     var value = await compute(
-        deserializeUserList, _result.data!.cast<Map<String, dynamic>>());
+      deserializeUserList,
+      _result.data!.cast<Map<String, dynamic>>(),
+    );
     return value;
   }
 
@@ -183,21 +193,23 @@ class _RestClient implements RestClient {
     final queryParameters = <String, dynamic>{};
     queryParameters.removeWhere((k, v) => v == null);
     final _headers = <String, dynamic>{};
-    final _data = <String, dynamic>{};
+    final Map<String, dynamic>? _data = null;
     final newOptions = newRequestOptions(options);
     newOptions.extra.addAll(_extra);
     newOptions.headers.addAll(_dio.options.headers);
     newOptions.headers.addAll(_headers);
     final _result = await _dio.fetch<List<dynamic>>(newOptions.copyWith(
-        method: 'GET',
-        baseUrl: baseUrl ?? _dio.options.baseUrl,
-        queryParameters: queryParameters,
-        path: '/usersNullable')
-      ..data = _data);
+      method: 'GET',
+      baseUrl: baseUrl ?? _dio.options.baseUrl,
+      queryParameters: queryParameters,
+      path: '/usersNullable',
+    )..data = _data);
     var value = _result.data == null
         ? null
         : await compute(
-            deserializeUserList, _result.data!.cast<Map<String, dynamic>>());
+            deserializeUserList,
+            _result.data!.cast<Map<String, dynamic>>(),
+          );
     return value;
   }
 
@@ -207,17 +219,17 @@ class _RestClient implements RestClient {
     final queryParameters = <String, dynamic>{};
     queryParameters.removeWhere((k, v) => v == null);
     final _headers = <String, dynamic>{};
-    final _data = <String, dynamic>{};
+    final Map<String, dynamic>? _data = null;
     final newOptions = newRequestOptions(options);
     newOptions.extra.addAll(_extra);
     newOptions.headers.addAll(_dio.options.headers);
     newOptions.headers.addAll(_headers);
     final _result = await _dio.fetch<Map<String, dynamic>>(newOptions.copyWith(
-        method: 'GET',
-        baseUrl: baseUrl ?? _dio.options.baseUrl,
-        queryParameters: queryParameters,
-        path: '/userByKey')
-      ..data = _data);
+      method: 'GET',
+      baseUrl: baseUrl ?? _dio.options.baseUrl,
+      queryParameters: queryParameters,
+      path: '/userByKey',
+    )..data = _data);
     var value = Map.fromEntries(await Future.wait(_result.data!.entries.map(
         (e) async => MapEntry(e.key,
             await compute(deserializeUser, e.value as Map<String, dynamic>)))));
@@ -230,17 +242,17 @@ class _RestClient implements RestClient {
     final queryParameters = <String, dynamic>{};
     queryParameters.removeWhere((k, v) => v == null);
     final _headers = <String, dynamic>{};
-    final _data = <String, dynamic>{};
+    final Map<String, dynamic>? _data = null;
     final newOptions = newRequestOptions(options);
     newOptions.extra.addAll(_extra);
     newOptions.headers.addAll(_dio.options.headers);
     newOptions.headers.addAll(_headers);
     final _result = await _dio.fetch<Map<String, dynamic>>(newOptions.copyWith(
-        method: 'GET',
-        baseUrl: baseUrl ?? _dio.options.baseUrl,
-        queryParameters: queryParameters,
-        path: '/userByKeyNullable')
-      ..data = _data);
+      method: 'GET',
+      baseUrl: baseUrl ?? _dio.options.baseUrl,
+      queryParameters: queryParameters,
+      path: '/userByKeyNullable',
+    )..data = _data);
     var value = _result.data == null
         ? null
         : Map.fromEntries(await Future.wait(_result.data!.entries.map(
@@ -257,17 +269,17 @@ class _RestClient implements RestClient {
     final queryParameters = <String, dynamic>{};
     queryParameters.removeWhere((k, v) => v == null);
     final _headers = <String, dynamic>{};
-    final _data = <String, dynamic>{};
+    final Map<String, dynamic>? _data = null;
     final newOptions = newRequestOptions(options);
     newOptions.extra.addAll(_extra);
     newOptions.headers.addAll(_dio.options.headers);
     newOptions.headers.addAll(_headers);
     final _result = await _dio.fetch<Map<String, dynamic>>(newOptions.copyWith(
-        method: 'GET',
-        baseUrl: baseUrl ?? _dio.options.baseUrl,
-        queryParameters: queryParameters,
-        path: '/usersByKey')
-      ..data = _data);
+      method: 'GET',
+      baseUrl: baseUrl ?? _dio.options.baseUrl,
+      queryParameters: queryParameters,
+      path: '/usersByKey',
+    )..data = _data);
     var value = Map.fromEntries(await Future.wait(_result.data!.entries.map(
         (e) async => MapEntry(
             e.key,
@@ -282,17 +294,17 @@ class _RestClient implements RestClient {
     final queryParameters = <String, dynamic>{};
     queryParameters.removeWhere((k, v) => v == null);
     final _headers = <String, dynamic>{};
-    final _data = <String, dynamic>{};
+    final Map<String, dynamic>? _data = null;
     final newOptions = newRequestOptions(options);
     newOptions.extra.addAll(_extra);
     newOptions.headers.addAll(_dio.options.headers);
     newOptions.headers.addAll(_headers);
     final _result = await _dio.fetch<Map<String, dynamic>>(newOptions.copyWith(
-        method: 'GET',
-        baseUrl: baseUrl ?? _dio.options.baseUrl,
-        queryParameters: queryParameters,
-        path: '/user')
-      ..data = _data);
+      method: 'GET',
+      baseUrl: baseUrl ?? _dio.options.baseUrl,
+      queryParameters: queryParameters,
+      path: '/user',
+    )..data = _data);
     final value = await compute(deserializeUser, _result.data!);
     return value;
   }
@@ -303,17 +315,17 @@ class _RestClient implements RestClient {
     final queryParameters = <String, dynamic>{};
     queryParameters.removeWhere((k, v) => v == null);
     final _headers = <String, dynamic>{};
-    final _data = <String, dynamic>{};
+    final Map<String, dynamic>? _data = null;
     final newOptions = newRequestOptions(options);
     newOptions.extra.addAll(_extra);
     newOptions.headers.addAll(_dio.options.headers);
     newOptions.headers.addAll(_headers);
-    final _result = await _dio.fetch<Map<String, dynamic>>(newOptions.copyWith(
-        method: 'GET',
-        baseUrl: baseUrl ?? _dio.options.baseUrl,
-        queryParameters: queryParameters,
-        path: '/userNullable')
-      ..data = _data);
+    final _result = await _dio.fetch<Map<String, dynamic>?>(newOptions.copyWith(
+      method: 'GET',
+      baseUrl: baseUrl ?? _dio.options.baseUrl,
+      queryParameters: queryParameters,
+      path: '/userNullable',
+    )..data = _data);
     final value = _result.data == null
         ? null
         : await compute(deserializeUser, _result.data!);
@@ -321,50 +333,57 @@ class _RestClient implements RestClient {
   }
 
   @override
-  Future<void> patchUser({required user, options}) async {
+  Future<void> patchUser({
+    required user,
+    options,
+  }) async {
     const _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{
       r'u': await compute(serializeUser, user)
     };
     queryParameters.removeWhere((k, v) => v == null);
     final _headers = <String, dynamic>{};
-    final _data = <String, dynamic>{};
+    final Map<String, dynamic>? _data = null;
     final newOptions = newRequestOptions(options);
     newOptions.extra.addAll(_extra);
     newOptions.headers.addAll(_dio.options.headers);
     newOptions.headers.addAll(_headers);
     await _dio.fetch<void>(newOptions.copyWith(
-        method: 'PATCH',
-        baseUrl: baseUrl ?? _dio.options.baseUrl,
-        queryParameters: queryParameters,
-        path: '/user/{user}')
-      ..data = _data);
-    return null;
+      method: 'PATCH',
+      baseUrl: baseUrl ?? _dio.options.baseUrl,
+      queryParameters: queryParameters,
+      path: '/user/{user}',
+    )..data = _data);
   }
 
   @override
-  Future<void> patchUserMap({required user, options}) async {
+  Future<void> patchUserMap({
+    required user,
+    options,
+  }) async {
     const _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     queryParameters.addAll(await compute(serializeUser, user));
     queryParameters.removeWhere((k, v) => v == null);
     final _headers = <String, dynamic>{};
-    final _data = <String, dynamic>{};
+    final Map<String, dynamic>? _data = null;
     final newOptions = newRequestOptions(options);
     newOptions.extra.addAll(_extra);
     newOptions.headers.addAll(_dio.options.headers);
     newOptions.headers.addAll(_headers);
     await _dio.fetch<void>(newOptions.copyWith(
-        method: 'PATCH',
-        baseUrl: baseUrl ?? _dio.options.baseUrl,
-        queryParameters: queryParameters,
-        path: '/userMap/{user}')
-      ..data = _data);
-    return null;
+      method: 'PATCH',
+      baseUrl: baseUrl ?? _dio.options.baseUrl,
+      queryParameters: queryParameters,
+      path: '/userMap/{user}',
+    )..data = _data);
   }
 
   @override
-  Future<void> postUsers({required users, options}) async {
+  Future<void> postUsers({
+    required users,
+    options,
+  }) async {
     const _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     queryParameters.removeWhere((k, v) => v == null);
@@ -375,16 +394,18 @@ class _RestClient implements RestClient {
     newOptions.headers.addAll(_dio.options.headers);
     newOptions.headers.addAll(_headers);
     await _dio.fetch<void>(newOptions.copyWith(
-        method: 'POST',
-        baseUrl: baseUrl ?? _dio.options.baseUrl,
-        queryParameters: queryParameters,
-        path: '/users')
-      ..data = _data);
-    return null;
+      method: 'POST',
+      baseUrl: baseUrl ?? _dio.options.baseUrl,
+      queryParameters: queryParameters,
+      path: '/users',
+    )..data = _data);
   }
 
   @override
-  Future<void> postUser({required user, options}) async {
+  Future<void> postUser({
+    required user,
+    options,
+  }) async {
     const _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     queryParameters.removeWhere((k, v) => v == null);
@@ -396,16 +417,18 @@ class _RestClient implements RestClient {
     newOptions.headers.addAll(_dio.options.headers);
     newOptions.headers.addAll(_headers);
     await _dio.fetch<void>(newOptions.copyWith(
-        method: 'POST',
-        baseUrl: baseUrl ?? _dio.options.baseUrl,
-        queryParameters: queryParameters,
-        path: '/user')
-      ..data = _data);
-    return null;
+      method: 'POST',
+      baseUrl: baseUrl ?? _dio.options.baseUrl,
+      queryParameters: queryParameters,
+      path: '/user',
+    )..data = _data);
   }
 
   @override
-  Future<void> postUserNullable({user, options}) async {
+  Future<void> postUserNullable({
+    user,
+    options,
+  }) async {
     const _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     queryParameters.removeWhere((k, v) => v == null);
@@ -419,37 +442,36 @@ class _RestClient implements RestClient {
     newOptions.headers.addAll(_dio.options.headers);
     newOptions.headers.addAll(_headers);
     await _dio.fetch<void>(newOptions.copyWith(
-        method: 'POST',
-        baseUrl: baseUrl ?? _dio.options.baseUrl,
-        queryParameters: queryParameters,
-        path: '/userNullable')
-      ..data = _data);
-    return null;
+      method: 'POST',
+      baseUrl: baseUrl ?? _dio.options.baseUrl,
+      queryParameters: queryParameters,
+      path: '/userNullable',
+    )..data = _data);
   }
 
-  RequestOptions newRequestOptions(Options? options) {
+  RequestOptions newRequestOptions(Object? options) {
     if (options is RequestOptions) {
       return options as RequestOptions;
     }
-    if (options == null) {
-      return RequestOptions(path: '');
+    if (options is Options) {
+      return RequestOptions(
+        method: options.method,
+        sendTimeout: options.sendTimeout,
+        receiveTimeout: options.receiveTimeout,
+        extra: options.extra,
+        headers: options.headers,
+        responseType: options.responseType,
+        contentType: options.contentType.toString(),
+        validateStatus: options.validateStatus,
+        receiveDataWhenStatusError: options.receiveDataWhenStatusError,
+        followRedirects: options.followRedirects,
+        maxRedirects: options.maxRedirects,
+        requestEncoder: options.requestEncoder,
+        responseDecoder: options.responseDecoder,
+        path: '',
+      );
     }
-    return RequestOptions(
-      method: options.method,
-      sendTimeout: options.sendTimeout,
-      receiveTimeout: options.receiveTimeout,
-      extra: options.extra,
-      headers: options.headers,
-      responseType: options.responseType,
-      contentType: options.contentType.toString(),
-      validateStatus: options.validateStatus,
-      receiveDataWhenStatusError: options.receiveDataWhenStatusError,
-      followRedirects: options.followRedirects,
-      maxRedirects: options.maxRedirects,
-      requestEncoder: options.requestEncoder,
-      responseDecoder: options.responseDecoder,
-      path: '',
-    );
+    return RequestOptions(path: '');
   }
 
   RequestOptions _setStreamType<T>(RequestOptions requestOptions) {
