@@ -1,30 +1,34 @@
 import 'dart:async';
 import 'dart:convert';
 import 'dart:typed_data';
-import 'package:dio/dio.dart';
+
 import 'package:dio/adapter.dart';
+import 'package:dio/dio.dart';
 
 class MockAdapter extends HttpClientAdapter {
-  static const String mockHost = "mockserver";
-  static const String mockBase = "http://$mockHost";
+  static const String mockHost = 'mockserver';
+  static const String mockBase = 'http://$mockHost';
   DefaultHttpClientAdapter _adapter = DefaultHttpClientAdapter();
 
   @override
-  Future<ResponseBody> fetch(RequestOptions options,
-      Stream<Uint8List>? requestStream, Future? cancelFuture) async {
-    Uri uri = options.uri;
+  Future<ResponseBody> fetch(
+    RequestOptions options,
+    Stream<Uint8List>? requestStream,
+    Future? cancelFuture,
+  ) async {
+    final uri = options.uri;
     if (uri.host == mockHost) {
       switch (uri.path) {
-        case "/tags":
+        case '/tags':
           return ResponseBody.fromString(
-            jsonEncode(["tag1", "tag2"]),
+            jsonEncode(['tag1', 'tag2']),
             200,
             headers: {
               Headers.contentTypeHeader: [Headers.jsonContentType],
             },
           );
         default:
-          return ResponseBody.fromString("", 404);
+          return ResponseBody.fromString('', 404);
       }
     }
     return _adapter.fetch(options, requestStream, cancelFuture);
